// src/types/quiz.ts
export interface QuizCreationRequest {
  title: string;
  description: string;
}

export interface Quiz {
  id: string;
  title: string;
  description: string;
  createdBy?: string;
  questionsCount?: number;
}

export interface QuestionCreationRequest {
  title: string;
  options: string[];
  correctAnswer: string;
  quizId: string;
}

export interface QuestionDto {
  title: string;
  options: string[];
}

export interface Question extends QuestionDto {
  id: string;
  correctAnswer?: string;
  quizId: string;
}

export interface QuestionResponse {
  answer: string;
}

export interface QuizStartRequest {
  quizId: string;
}

export interface AIQuizStartRequest {
  topic: string;
  size: number;
  level: string;
}

export interface WebSocketMessage<T> {
  type: string;
  payload: T;
}

export interface QuizResult {
  resultID?: string; // UUID generated by backend
  quizId: string;
  username: string;
  totalQuestions: number;
  correctAnswers: number;
  answers: Record<number, string>; // Map of question number to answer
  startedAt: string;
  endedAt: string;
}